Only scrapes terms that haven't been scraped already

+++ linkedin.rb
@@ -17,11 +17,13 @@ end
 
 # Go through all terms and scrape
 file.each do |term|
-  l = LinkedinData.new(term["Search Term"], term["Degrees"].to_i)
-  scrapeout = l.getData
-  filename = resultsdir+"/"+term["Search Term"].gsub(" ", "_").gsub("/", "-")+".json"
-  File.write(filename, scrapeout)
-  File.write(filename.gsub(".json", ".csv"), `json2csv '#{filename}'`)
+  if !File.exist?(term["Search Term"].gsub(" ", "_").gsub("/", "-")+".json")
+    l = LinkedinData.new(term["Search Term"], term["Degrees"].to_i)
+    scrapeout = l.getData
+    filename = resultsdir+"/"+term["Search Term"].gsub(" ", "_").gsub("/", "-")+".json"
+    File.write(filename, scrapeout)
+    File.write(filename.gsub(".json", ".csv"), `json2csv '#{filename}'`)
+  end
 end
 
 # Move the pictures directory to the results folder
